Add files via upload
Using the ESP8266, a web page is created on a private IP address. This Web page can be opened on multiple devices connected to the same network. DHT Sensors attached to the ESP8266 provide temperature and humidity data that can be seen on the web page. The digital button on the web page can be used to switch on and off the LED ON ESP8266. A graphic on the web page will show the status of the button on every device. An OLED display is used to show the IP address to access web page generated by ESP8266.